UPBGE: Fix orthographic billboard and halo.
Previously in both BFBGE and UPBGE the transformation computing for
billboard and halo material was only for perspective camera because it
needed the distance between the camera and the object.

The distance exists when the camera is an object but it is missing when
it is the override camera. Anyway orthographic camera doesn't have distance
to the objects.

To avoid using a wrong distance in RAS_OpenGLRasterizer::GetTransform we
add a condition if the camera is orthographic, if it is the case the
left vector (which was before the direction to the camera) is replaced
by the normalized Z orientation of the camera.