UPBGE: Replace STR_String by std::string.
Using std::string instead of STR_String is benefic for the following reasons:
- it requests less works to use and already existing standart class
- std::string can be compiler side optimized
- std::string supports more functions than STR_String

All places using STR_String are replaced by std::string, but old places using
const char * or char * are now trying to use std::string. In some cases is
not possible for example the blender file and path utilities.

The major refactor to support std::string is for python attributes class
KX_PYATTRIBUTE_DEF. Previously we know that the attributes list ends when
the name of an attribute was to NULL. The issue with std::string is that
is not a pointer and NULL could not be anymore used. To solve this issue
a default attribut with an empty name is added: KX_PYATTRIBUTE_NULL. Its
name length is checked to know if it's the last attribut.

Even if std::string has more function than STR_String, this class lack of
format and upper functions. To compensate this lack the boost functions
are used.

As STR_String is still used in blender by the GHOST part, a constructor
doing the conversion from std::string to STR_String is written.