UPBGE: Fix object friction setting convertion from defautl material.
Previously in 687df77ec09819d341db8f6b17c5c7ad39c92119 the physics settings
from the materail were converted to the object. But this was done only for
valid material and not default material.

To solve this issue the friction is set to 0.5 (the only non-zero value of
the default material) if none material did a convertion.

Fix issue #312.