UPBGE: Fix normal update in BL_MeshDeformer::RecalcNormals.
Previously the normal update was iterating the vertex by doing an iteration
of 3 by 3 on the indices. But this assume that the mesh is only using polygones
and not lines like the wire material allow.

As excepted the usage of lines raise segmentation fault because of wrong
memory access.

As lines are depending of polygones but doesn't use the same indices format as
polygones, to solve the normal computation we have to iterate on the mesh
polygones and find the normal for each polygons.

The normal update is a little bit slower than before but more adaptative as
it can do polygones and lines.

Fix issue #315.