UPBGE: Don't compare string adresse instead of content.
Previously when the returned value of CValue::GetText() was a reference,
CBoolValue was returning the value of a string allocated in heap. This
allowed to later write comparaison with only adresse like:

&boolval->GetText() == &CBoolValue::sTrueString

But now as the GetText function doesn't return a reference, the string
is copied and the adresse comparaison is broken.
To solve this issue we just back to a string content comparaison, which
is in every case more safer.