UPBGE: Fix lib loaded realtime cube map.
Previously when the user lib loaded a scene containing a realtime cube map.
The realtime wasn't doing a render.
Indeed the realtime cube map wasn't merged in the cube map manager of the
final scene.

To fix this issue a function named Merge is added in KX_CubeMapManager.
This function move the realtime cube maps of the merged scene to the the
final scene.

In the same way the texture have to be valid when the merging is made, but
this wasn't the case. Indeed the texture were initialized in
KX_BlenderMaterial::OnConstruction(). This function is wanted for the material
conversion because it depends of the scene used. But in the case of the
textures it is useless. The texture are not scene dependent and theirs
initialization is moved in KX_BlenderMaterial constructor.