UPBGE: Fix ImageRender image pixel buffer copy.
Previously with the introduction of HDR in 732d176d09b6c2497232bd3ecfced9042b66deff,
the image internal format needed to be variable depending of the HDR level.
To did this, ImageBase::m_format was added, but its usage was wrong due to an
ambiguity between image format and image internal format.

The image format was replaced by the internal image format and caused invalid pixels
reading. To solve this issue the place using the image format back to GL_RGBA
and GL_RGB. And the place really using the image internal format are now using
the variable m_internalFormat to be explicit.

Bug reported by sdfgeoff: #311.