UPBGE: Fix filter reading and writing to same off screen.
Previouly when only one filter where rendered with an overlay scene some artifacts
were noticed. The bug was caused by the fact that when only one filter was rendered
the targeted off screen was the same as the input off screen.

To avoid this we need to read from a different off screen that the one we are
drawing to.
To achieve this we can copy the data to a second off screen before rendering the
filters if we detect that there's only on filter. This method is simple but expensive.
The alternative is to target always a different off screen. It's the same behaviour
as for eye off screen:
If we are in multisample we target the same off screen than input because we know
that filters will do a blit/copy. If it the last scene optimize by chosing a non
multisample off screen.
Else in case of non-multisample, choose the next render off screen using NextRenderOffScreen.