UPBGE: Use names table for scene callbacks list.
Previously the three callbacks in KX_Scene were using a different function
for set and get, but the onyl difference between these functions was the
enumeration used to access to KX_Scene::m_drawCallbacks.

To simplificate the code only one getter and setter are used but they both
look in a table to find the coorespondent enumeration for the current attribut
name.