UPBGE: Fix absolute active mesh color layer index.
Previously when using CustomData_get_active_layer_index the value returned
was depending of all other layer before, so if we are looking for a color
layer and that there's 8 uv layer, the color layer index will be 9 (including
the 1 gap).
But the color layer index must be relative because we use the same value to
get the proper color in the vertex.

To have relative instead of absolute index, the function CustomData_get_active_layer
is used.

Reported by alabd14313 and youle #341.