UPBGE: Implement nodal render system.
Previously the render was proceed in a RAS_MaterialBucket function iterating
over every RAS_DisplayArrayBucket and calling their render function and finally
RAS_MaterialBucket::RenderMeshSlot was called. But the performance issue noticed
with this system is that if you don't have any display array bucket you will
even bind/unbind the material and the same will happen for display array bucket
if there's no mesh slots. To avoid this issue the code was counting recursively
the number of mesh slot for a material ect…, and these values were cached as
they are expensive to compute.

With the coming of mesh batching a much bigger problem appeared. Previously the
sorted render for transparency meshes was containing only meshes not using
geometry instancing, but it should also avoid sort mesh slots using mesh
batching. Here the problem is that there's no way to extract the sortable mesh
slots from the others as the system is considering using only one category of
mesh slots.

To fix the both issue a node system was developed. This node system consists of
storing the function and data to use when rendering.
Each node contains:
	- a data to pass to the functions
	- a bind function
	- an unbind function

The bind function is called with the node's data at binding and the unbind
function as unbinding.
Technically a class RAS_BaseNode is defined using templates: InfoType, Args.
With these datas the function signature is:
void (InfoType *, const Args&).
The node is constructed passing an InfoType instance and two function pointers.
In addition the class has a Bind(const Args&) and Unbind(const Args&).

The most simple system replaced with nodes is the unsorted render. For this we
have to use a material node containing a list of display array bucket nodes
which are itself containing mesh slot nodes. A new type of node can be defined:
RAS_DownwardNode, it design a node iterable from the root of the tree to the
leafs. This class use a new template ChildType which is the child node type used
in a list contained in RAS_DownwardNode.
During the render the Execute function of a downward node is called. This
Execute function call the binding function then iterating over all children
nodes and calling the same Execute function, then at the end call its unbinding
function.

Now that we know how to render downward nodes, we need to define a way to
generate the tree of downward nodes. To doing this a function named GenerateTree
is added in the classes using nodes: RAS_MaterialBucket, RAS_DisplayArrayBucket
and RAS_MeshSlots. This function takes as arguments the root nodes, and an other
kind of root node.

The sorted render behave a way differently than the unsorted render, in this
render we have to break the priority make of the parent nodes and use the
priority of the leafs node order. In our case, the leafs are mesh slot nodes and
they are ordered by the distance between the mesh slot and the camera.
During the render we have to iterate on a list of leaf nodes. At the first
iteration we bind the parents of the leaf from the top to the bottom of the
tree. At the next iteration we check if the parent of the previous leaf is
different than the parent of the current leaf, if not, nothing is done, else the
parent of the previous leaf is unbound and the current leaf parent is bound,
then we call the same check for the the parent of parent.

To do a such behavior, the node just have to know its parent node, a new class
answering this goal is added: RAS_UpwardNode. The generation of the upward node
tree is also a bit different. Firstly we use the same function GenerateTree,
secondly the node of material, display array bucket or mesh slot is setting
their parent from the root node passed in GenerateTree function.

As we have to know the previous parent node, the render of the upward can't be
done in a function defined from RAS_UpwardNode. The previous parent is stored in
a new class named RAS_UpwardNodeIterator. This class contains a recursive
function NextNode which manage the check and switches of parent nodes.

All the arguments used for the binding and unbinding functions are packed into
the class RAS_RenderNodeArguments, this class contains the camera
transformation, the rasterizer pointer, a boolean for sorting render and a
boolean for shader override.

All the sources about render nodes are put in the directory Rasterizer/Node. The
node definition corresponding to real used type are in RAS_RenderNode.h.
No performance regression were noticed with the render node system currently.