UPBGE: Fix keyboard sensor typed text.
Previously in SCA_KeyboardSensor::LogKeystrokes the typed string was appended
to the property string if the user wanted it.
But with the last commit replacing reference to regular STR_String in function
GetText() and GetName(): using const char* val = prop->GetText().ReadPtr()
is dangerous as we want to have a pointer of a string which is temporary.

To solve this we just make the string local instead of temporary.