UPBGE: Implement cube map lod update.
Previously in the cube map render the lod wasn't proceed and use the last lod
state from the previous render.
Lod can be usefull to reduce render time during cube map with the combinaison
of a different lod sistance factor for the cube maps.

To solve this lack, KX_CubeMap store the lod distance factor under the name
of m_lodDistanceFactor which is exposed in python as: KX_CubeMap.lodDistanceFactor.
In KX_CubeMapManager::RenderCubeMap this value is set in the camera used to
render. When rendering each face a call to KX_Scene::UpdateObjectLods is made.
This is better than calling the lod update without the culling of the face because
at the end the faces are normally showing every objects but with the advantages
of oclusion culling and frustum clipping.