UPBGE: Refactor and allow switching more than two lod levels.
Previously the lod manager update function was only managing the case where the
current distance was correct for the next level or for the previous level. But
if the distance changed of a big value and requested to change more than one lod
level, the manager failed selecting the proper level.

To increase the readability of the code and fix this issue, an iterator class is
used. The pinciple in simple: When you compare if the iterator is smaller than
the current distance you increment it. Else if the iterator is bigger than the
current distance you decrement it. When the two condition failed is that you
found the proper level, to get its index you defered the iterator.

In the same way m_previousLodLevel which was useless is removed and replaced by
m_currentLodLevel.
Also KX_LodManager::GetLevel(scene, distance) returning the KX_LodLevel instead
of it's index when the lod changed, else it returns NULL.