UPBGE: Expose layer in KX_GameObject.
The object layer is used for culling stage in two case: shadows and realtime cube
map render.
But the user was able to change the layer value of the cube map or the light only.
In some case when the user want to avoid render shadow for only one object without
setting the object invisble, he could change the object layer to fail check with
light layer.

To do this the layer of each object is exposed under:
KX_GameObject.layer