UPBGE: Fix parallax mapping artifacts.
Previously artifacts were noticed in the area where the shader change of step, after
some test it appears that using a mipmapped texture to compute the parallax uv is
unsafe.

Replacing a call to texture2D by a call to textureLod(..., 0) avoid using any mipmap
and resolve the artifacts. Even if we don't use mipmapping to compute the parallax uv,
all the texture using this uv can keep using mipmapping.

In the same time the unused function parallax_scale is removed.