UPBGE: Avoid bind scene of physics environment during scene conversion.
Binding the scene under conversion and its physics environment was introduce with
the implementation of the python components. But as we moved the python initialization
of the components out of the conversion function, binding the scene and its physics
environment is useless and dangerous becasue there's no back up to the scene bound
previously.

To solve this issue the two line binding are removed.

Reported by ondredeniso in #349.