UPBGE: Fix lod update with zero distance factor.
When the distance factor is to zero then all the object to camera distance are to zero.

Previously in this case in KX_LodManager::GetLevel we tested with it <= distance,
it return that 0 is equal to minimum of the first lod level (the one containing the
original mesh).
At the next iteration it <= distance failed because its maximum is not 0, so we go to
it >= distance and it says true because 0 is equal to the minimum of the second lod
level.
Then an infinite loop is created. To solve it we simply replce the it >= by it >
Only one of the bound (minimum or maximum) is allowed to use more than its operator
the = operator.