UPBGE: Fix font "Text" property conversion.
Previously the user was able to set a the "Text" property of a font object to an
other type than string, e.g int, float, boolean. But in this case the string
contained in the font object must be able to success the conversion. Else a default
value is used.

This behaviour was made before the switch to std::string, but with the equivalent
conversion functions for std::string, exceptions were raised instead returning a
default value.

To avoid handling exception the function std::stof and std::stoi are replaced by
the use of the operator >> of a std::stringstream. The string of the font object
is pushed to the stream then, conresponding of the property types the stream push
to an int or a flot or a boolean. The state of the stream is checked at the end
of the conversion function. If its state notices that the conversion failed, an
error message is printed.