UPBGE: Remove useless code from Expression/ part.
A lot of functions from the Expression/ part were not used and wrapped into other unused
functions.
The most sensible part is the COperator2Epr m_cached_calculate value. This value was
allowing to update the result only when the expression used were modified, but the function
SetModified() which set the expression modified was never called.