UPBGE: Remove BLI_asset in CExpression and CValue destructor.
Previously an BLI_assert was called in the destructor to prevent the case
were a CValue was not delete by Release() function.
But in some case a call to Release is not possible because the CValue
is not the base class but one of the final classes. The class BL_Texture
is under this case. BL_Texture is based on RAS_Texture and CValue but the
delete is from RAS_IPolyMaterial with the RAS_Texture side.

To avoid useless assert, they are now removed.