UPBGE: Refactor debug shape render.
Previously all the debug shapes (line, circle and box) were using a common
struct: OglDebugShape, but the need of other debug shapes like trapezoidal
shape will put in the following situation: The smaller debug shape type will
use the same memory of the bigger debug shape type. Also as these type are
common we need first to store a type variable in it but the worse is that
during the render we have to check for every shape it's type.

To avoid such issue each debug shape use its own type. The only common point
between all the type is the color. To avoid defining the color everywhere all
the type are sub struct of DebugShape. In the same time it allow use to rename
the struct attribut to more explicit names, m_param and m_param2 are now
removed.

As the debug shapes are no longer a common type we need a different list to
store lines, circle, and boxes. This is the goal of struct SceneDebugShape.
This struct replace the std::vector previously used as value in m_debugShapes.

The current debug shape type "Box" is not really a box because it doesn't
contain the 8 vertices but only the minimum and maximum corner as an AABB.
To avoid name conflict, they are renamed "Aabb".