UPBGE: Implement filter mipmapping render texture.
Previously the rendered texture sent to the filters shaders was without
mipmapping. In some case like fast blur the user could need filtering the
rendered texture based on mipmap levels.

As generating mipmaps is quite expensive the generation is optional and
requested by setting in python KX_2DFilter.mipmap to True.

Internally the texture should be allowed to filter basing on its mipmaps
but not all the time, so the function RAS_IRasterizer::MipmapOffScreenTexture
set the texture filtering to mipmap mode and generate the mipmap. After
the use of the rendered texture, the texture  must back to no filter mode
thanks to calling the function RAS_IRasterizer::UnmipmapOffScreenTexture.

The GPU function GPU_texture_filter_mode accept the new argument "mipmap"
to set the min filtering to GL_LINEAR_MIPMAP_LINEAR.