UPBGE: Fix parallax mapping backface render.
Previously when a face of a mesh using a material with paralla mapping, the
vector used to sample the step of the parallax was wrong in Y axis when the
render was made from the back.
A lot of try to use a common behaviour to be non-face dependent was made, but
the most simple is to simply use a check with gl_FrontFacing and revert the
sample vector Y axis.

Fix issue #367.