UPBGE: Fix global dict double decref and sub modules incref.
Previously the global dict (stored in bge.logic) was add to the next bge.logic
module in case of game restart in LA_Launcher::ExitEngine.
But this action is useless and dangerous. First it is dangerous because the
globalDict in GPG_ghost could be decref an then delete before calling
ExitEngine. Secondly this is useless because the global dict is reset at each
call of setupGamePython.

So the decref of global dict is removed in ExitEngine.

In the same time an extra incref was procceed in initBGE for sub modules.
The call to PyModule_AddObject was not incref but the call to PyDict_SetItemString
was incref. So no need of manually incref. This incref is removed ant the
sub modules add functions are moved into addSubModule(…).

Orignal works of youle.

Fix issue #350.