UPBGE: Fix material name check without name.
Previously in some sensors like collision, ray and constraints sensors
a material check was possible. As all material name have to start with
the "MA" prefix because video texture is based on this to check that it
have a material or a texture.

But in these sensors the names without prefix were used, so the prefix
where remove by using char + 2, at the std::string translation is became
std::string(name, 2). But this isn't working if the name is empty.
This situation appears for the default material.

The easiest solution was to make sure that every material name start with
the "MA" prefix.

Fix issue #357.