UPBGE: Add an option to debug draw lights shadow frustum.
Debugging the shadow frustum box is useful to optimize the render
of a shadow by restrict the render to a limited group of objects.

The option used for debugging the shadow frustum in viewport is extended
to the spot and this option is also used in the game engine.

To control if the debug in the viewport is also wanted into the game,
the debug option "Show Shadow Frustum" is added in the Debug panel.
If this option is to Allow (and the shadow box option is enabled per
lamp) or Force, each lamp objects are drawing a frustum box similar to
the camera frustum. This is proceeded in DrawDebugShadowFrustum reusing
the debug draw function DrawCameraFrustum as shadow frustums are
cameras frustum.

The debug frustum value per lamp are stored into KX_LightObject and
modified by the getter setter [Get/Set]ShadowFrustum.