UPBGE: Fix normal transformation in batching and in KX_MeshProxy.transform.
Previously the matrix used to transform the vertices normal in both batching
and in KX_MeshProxy, was an inverse transposed with null translation matrix
from the original transformation matrix.
Avoiding a translate is logical, but inversing the matrix and then transpose
isn't very logic. These operations caused to rotate the normal from the geometry
when the vertices are transformed. What we need is just to rotate the normals
as the vertices.
Removing the inverse and transpose produce correct results.

Fix issue #365.