UPBGE: Fix mesh batching increasing merge time.
Previously the merging time of the same mesh multiple time in a batching
group was far to be constant time. The reason was the use of std::vector::reserve
and push_back. Even if the reserve call should avoid multiple allocation
the both function are a lot slower than the combinaison of std::vector::resize
and index access and assignement.

The difference is of about 1ms now vs 13ms before.

After some benchmarks, the reserve and push_back usage is about 2 to 3
times slower than a resize with index assignement.

Now the time is less increasing but still not constant, this is due to the
vertex cache pointer regeneration from function UpdateCache.
The cost of this function could be minimized by calling it only at the end
of multiple merging.