UPBGE: Fix include from blenderplayer python main loop script.
Previously the python main loop script was just run as it. So the include
depending on other files located in the same directory than the script
were failing the import.

To fix this issue we have to add the script directory into the python
system path list.
To realize this, a new function LA_Launcher::RunPythonMainLoop is added,
its goal is to run the python main loop and in the LA_PlayerLauncher
override, add the include path (if we use a python script passed from
commad line) before run the python main.

In the same way the complexity of char *(*) used in LA_Launcher and
KX_PythonMain is removed by using simple std::string.