UPBGE: Fix upbge version check.
The version is composed of three digits, e.g 0.1.5.
But the two first digits are stored in the version number and the last
in the sub version number.
But previously MAIN_VERSION_UPBGE_ATLEAST was used with the version for
the first digit and the subversion for the last digits. This as breaking
compaitibility upper 0.1.0.

To fix this issue the two first digit are passed as first arguments of
MAIN_VERSION_UPBGE_ATLEAST.