UPBGE: Fix compilation errors on MSVC 2013 and C2664 error.
The issue in MSVC 2013 compilation was that the compiler wasn't supporting the
default constructor specified with using keyword. To solve this we have to set
the constructor in RAS_RenderNode.h.

In the second time the std::function in MSVC wasn't supporting NULL for null
function pointer, here it requests to use nullptr.
In the third time the class function pointer can't be passed as it, it requests
to call std::mem_fn to return a proper callable function.