UPBGE: Add lod distacne factor in UI.
The lod distance factor is supported by objects, cameras and environment
maps. But its vaue was changeable only from python and not from the UI.

To solve this lack the distance factor is exposed in the three places.
These value are stored in the blender struct under the name of lodfactor.